+++ example/android/app/src/main/kotlin/com/tieudan/k_responsive_example/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.tieudan.k_responsive_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}