+++ ios/Classes/KResponsivePlugin.h
@@ -1,0 +1,4 @@
+#import <Flutter/Flutter.h>
+
+@interface KResponsivePlugin : NSObject<FlutterPlugin>
+@end