+++ ios/Classes/SwiftKResponsivePlugin.swift
@@ -1,0 +1,14 @@
+import Flutter
+import UIKit
+
+public class SwiftKResponsivePlugin: NSObject, FlutterPlugin {
+  public static func register(with registrar: FlutterPluginRegistrar) {
+    let channel = FlutterMethodChannel(name: "k_responsive", binaryMessenger: registrar.messenger())
+    let instance = SwiftKResponsivePlugin()
+    registrar.addMethodCallDelegate(instance, channel: channel)
+  }
+
+  public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
+    result("iOS " + UIDevice.current.systemVersion)
+  }
+}