+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+# Change Log
+
+## 0.1.0 Break Change
+
+- Update Readme
+
+- Update syntax old
+
+- Add method update when you only responsive for a device eg: tablet and something else is defaultValue
+
+- Remove all properties about MediaQuery in KContext to avoid conflic with difference extension
+
+## 0.0.3
+
+- fix `getType` change compare < to <=
+
+## 0.0.2
+
+- upgrade document
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+- TODO: Describe initial release